+++ src/course.cpp
@@ -1,6 +1,6 @@
 #include "course.hpp"
 
-Course::Course(const std::uint8_t id, const std::string title,
+Course::Course(const std::uint16_t id, const std::string title,
                const std::uint8_t capsize)
     : _id(id), title(title), capsize(capsize) {}
 
@@ -14,4 +14,4 @@ bool Course::register_student(Student& student) {
     return true;
 }
 
-std::uint8_t Course::id() const { return this->_id; }
+std::uint16_t Course::id() const { return this->_id; }
