+++ src/course.hpp
@@ -6,16 +6,16 @@
 
 class Course {
 public:
-    Course(const std::uint8_t, const std::string, const std::uint8_t capsize);
+    Course(const std::uint16_t, const std::string, const std::uint8_t capsize);
 
     bool register_student(Student&);
 
-    std::uint8_t id() const;
+    std::uint16_t id() const;
 
 private:
-    const std::uint8_t _id;
+    const std::uint16_t _id;
     const std::string title;
     const std::uint8_t capsize;
 
-    std::unordered_map<std::uint8_t, Student&> students;
+    std::unordered_map<std::uint32_t, Student&> students;
 };
