+++ src/student.cpp
@@ -1,6 +1,6 @@
 #include "student.hpp"
 
-Student::Student(const std::uint8_t id, const std::string name)
+Student::Student(const std::uint32_t id, const std::string name)
     : _id(id), _name(name) {}
 
-std::uint8_t Student::id() const { return this->_id; }
+std::uint32_t Student::id() const { return this->_id; }
